fix(global): fix critique button auth verif; create linked page dont exist; replace alert to toast

+++ src/actions/delete-test-actions.tsx
@@ -11,6 +11,19 @@ export async function deleteTest(projectId: string, testId: string) {
         return { error: "Vous devez être connecté pour créer un test" };
     }
 
+    const isMember = await prisma.member.findUnique({
+        where: {
+            userId_organizationId: {
+                userId: user.id,
+                organizationId: projectId,
+            },
+        },
+    });
+
+    if (!isMember) {
+        return { error: "Vous n'avez pas les permissions pour supprimer ce test" };
+    }
+
     await prisma.test.delete({
         where: {
             id: testId,
